remove typescript from db manager
